+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to
 
 - ♻️(addon) improve Outlook add-on: i18n support, feedback link, smarter link
 
+### Fixed
+
+- 🐛(frontend) fix noise reduction left-channel-only audio
+
 ## [1.19.0] - 2026-06-04
 
 ### Added